chore: implemented current user

+++ backend/src/services/recipes.js
@@ -1,19 +1,17 @@
 import { Recipe } from '../db/models/recipe.js'
 
 // Create a recipe service function ===========================================
-export async function createRecipe({
-  title,
-  author,
-  ingredientList,
-  imageURL,
-}) {
-  const recipe = new Recipe({ title, author, ingredientList, imageURL })
+export async function createRecipe(
+  userId,
+  { title, ingredientList, imageURL },
+) {
+  const recipe = new Recipe({ title, author: userId, ingredientList, imageURL })
   return await recipe.save()
 } // End create recipe
 
 // List all the available recipes in the database =============================
 //=============================================================================
-// It is a helper/common function==============================================
+// It is a helper/common function================
 async function listRecipes(
   query = {},
   { sortBy = 'createdAt', sortOrder = 'descending' } = {},
@@ -25,12 +23,10 @@ async function listRecipes(
 export async function listAllRecipes(options) {
   return await listRecipes({}, options)
 }
-
 // List recipes by an author ==================================================
 export async function listRecipesByAuthor(author, options) {
   return await listRecipes({ author }, options)
 }
-
 // List all recipes by tags ===================================================
 export async function listRecipesByTag(tags, options) {
   return await listRecipes({ tags }, options)
@@ -45,16 +41,17 @@ export async function getRecipeById(recipeId) {
 // Update the recipe given a recipe ID ========================================
 export async function updateRecipe(
   recipeId,
-  { title, author, ingredientList, imageURL, tags },
+  userId,
+  { title, ingredientList, imageURL, tags },
 ) {
   return await Recipe.findOneAndUpdate(
-    { _id: recipeId },
-    { $set: { title, author, ingredientList, imageURL, tags } },
+    { _id: recipeId, author: userId },
+    { $set: { title, ingredientList, imageURL, tags } },
     { new: true },
   )
 }
 
 // Delete a post given a recipe ID ============================================
-export async function deleteRecipe(recipeId) {
-  return await Recipe.deleteOne({ _id: recipeId })
+export async function deleteRecipe(userId, recipeId) {
+  return await Recipe.deleteOne({ _id: recipeId, author: userId })
 }
